Begin refactoring of the IR interpreter
Start of the move away from a visitor pattern to an interpreter pattern.

Hopefully most of the functionality for nodes has been transferred
across, but this is all currently untested due to weird incompatibilities
between this branch and my environment. The changes made will need to be
picked onto a new branch off of upstream Graal.